pyproject.toml
[project]
name = "wiki-agent-benchmark"
version = "1.0.0"
description = "MIPROv2 + ACE Research Pipeline for HotpotQA with LLMControls API"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "wikipedia>=1.4.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "dspy-ai>=2.5.0",
    "datasets>=2.19.0",
    "scikit-learn>=1.4.0",
    "nest-asyncio>=1.6.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
run-pipeline = "scripts.run_pipeline:main"
